faceverif/requirements.txt
tensorflow>=2.7.0
numpy>=1.21.2
opencv-python>=4.5.3.56
matplotlib>=3.5.1
Pillow>=9.0.0
Flask>=2.0.2
requests>=2.26.0
pandas>=1.3.5
deepface>=0.0.70
tqdm>=4.62.3
keras>=2.7.0
mtcnn>=0.1.0
gunicorn
